// dashboard/apps/web/modules/layouts/main/nav/types.ts
import { ListItemButtonProps } from '@mui/material/ListItemButton';
import { StackProps } from '@mui/material/Stack';

// ----------------------------------------------------------------------

export type NavItemStateProps = {
  open?: boolean;
  active?: boolean;
  subItem?: boolean;
  hasChild?: boolean;
  externalLink?: boolean;
};

export type NavItemBaseProps = {
  title: string;
  path: string;
  icon?: React.ReactElement;
  type?: 'simple' | 'subheader';
  children?: NavItemSubheaderProps[] | NavItemBaseProps[];
};

export type NavItemSubheaderProps = {
  subheader: string;
  items: {
    title: string;
    path: string;
  }[];
};

export type NavItemProps = ListItemButtonProps &
  NavItemBaseProps &
  NavItemStateProps;

export type NavListProps = {
  data: NavItemBaseProps;
};

export type NavSubListProps = StackProps & {
  data: NavItemBaseProps[];
  subheader: string;
};

export type NavProps = {
  data: NavItemBaseProps[];
};